pyproject.toml
[project]
name = "multi-aghent-ai-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "dotenv>=0.9.9",
    "openai-agents>=0.0.19",
]
